Refactor validation rule for capacity
